refactor(ui): consolidate OIDC trust-policy create/edit into one mode-prop form
