+++ services/dashboard-ui/client/components/oidc-trust-policies/CreateOIDCTrustPolicy/index.ts
@@ -1,7 +1,1 @@
 export { CreateOIDCTrustPolicyButton } from './CreateOIDCTrustPolicyContainer'
-export {
-  CreateOIDCTrustPolicyModal,
-  hasSubCondition,
-  type ClaimCondition,
-  type OIDCTrustPolicyFormInput,
-} from './CreateOIDCTrustPolicy'
