+++ services/dashboard-ui/client/components/oidc-trust-policies/EditOIDCTrustPolicy/index.ts
@@ -1,2 +1,1 @@
 export { EditOIDCTrustPolicyButton } from './EditOIDCTrustPolicyContainer'
-export { EditOIDCTrustPolicyModal } from './EditOIDCTrustPolicy'
