+++ services/dashboard-ui/client/components/oidc-trust-policies/OIDCTrustPolicyForm/index.ts
@@ -1,0 +1,1 @@
+export { OIDCTrustPolicyFormModal } from './OIDCTrustPolicyForm'
